fixed slight issues on the confirmation pg and vaccine pg
For confirm.js: edited the user message to display user location. Also changed the direct link to book another appointment as to not have to type in user info again.

Timebooking.js: fixed an error associated with not adding a key value when using the object.map() function

Vaccines.js:changed how user msg is displayed